PC Assembly Lab
PC Assembly Lab
Date: 30/10/2025

We attended a PC assembly lab session accompanied by our lecturer, Dr. Aryati on 30th October. In this session, we were split into groups and basically we learned how to assemble and disassemble a PC physically. The most significant takeaway was bridging the gap between theory and practical. We always learn about the components of a system unit from textbooks and Internet videos, but this opportunity to physically handle and assemble them gave us a completely different level of understanding.

Besides, I found it challenging on the cable management because we need to be patient and pay attention to the manual in order to connect the wires to the correct pins on the motherboard. It taught me the importance of documentation and double-checking connections.

Overall, it was a memorable experience and I am truly grateful to Dr. Aryati for giving us this opportunity to learn and gain hands-on experience.